backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[argon2]
pyotp
vaderSentiment
python-socketio
slowapi
python-multipart
httpx
validators
beautifulsoup4
python-whois
qrcode
python-dotenv
